+++ grc/partnership_details/forms.py
@@ -1,6 +1,6 @@
 from flask_wtf import FlaskForm
 from grc.business_logic.data_structures.partnership_details_data import CurrentlyInAPartnershipEnum
-from grc.partnership_details.constants import PartnershipDetailsConstants as c
+from grc.business_logic.constants.partnership_details import PartnershipDetailsConstants as c
 from grc.lazy.lazy_fields import LazyRadioField
 from grc.lazy.lazy_form_custom_validators import LazyDataRequired
 from wtforms import StringField
